refactor(web): migrate mongo-query-generator to feature folder
