+++ apps/web/src/tools/mongo-query-generator/index.ts
@@ -1,0 +1,5 @@
+import type { ToolModule } from "@/tools/types";
+
+import { MongoQueryGenerator } from "./ui";
+
+export const tool: ToolModule = { id: "mongo-query-generator", Component: MongoQueryGenerator };
