+++ apps/web/src/tools/mongo-query-generator/messages.ts
@@ -1,0 +1,17 @@
+import type { LocaleMessages } from "@/tools/types";
+
+export const messages: LocaleMessages = {
+  en: {
+    Tools: {
+      "mongo-query-generator": {
+        title: "Filter → Mongo Query",
+        description: "Generate MongoDB queries from filter metadata.",
+      },
+    },
+  },
+  "zh-TW": {
+    Tools: {
+      "mongo-query-generator": { title: "篩選 → Mongo 查詢", description: "從篩選 metadata 產生 MongoDB 查詢。" },
+    },
+  },
+};
